requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.0.350
openai==1.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0
jinja2==3.1.2
